Added Zombie class and functions
I added Zombie.hpp, Zombie.cpp, main.cpp, and Makefile. The Zombie class has a constructor, a destructor,
a getter and a setter, and the announce function. All of the functions are functional.
Maybe look into adding a constructor that has the ability to set the name when calling it?
Start work on the next part which is adding the randomChump stuff